# backend/src/main/resources/openapi/documentation.yaml
openapi: 3.0.3
info:
  title: Piston Control / AgriTech API
  version: 1.0.0
  description: |
    API documentation exposed by the backend container.
servers:
  - url: http://localhost:8080
    description: Local Docker backend
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
  /auth/login:
    post:
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verify email code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                code:
                  type: string
              required: [userId, code]
      responses:
        '200':
          description: Verification handled
  /auth/resend-code:
    post:
      tags: [auth]
      summary: Resend email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
              required: [userId]
      responses:
        '200':
          description: Code resent
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
      required: [firstName, lastName, email, phoneNumber, password]
...